requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
babel==2.16.0
bcrypt==4.0.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.4
cryptography==43.0.3
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd==1.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.1
passlib==1.7.4
pillow==11.0.0
pluggy==1.5.0
plyer==2.1.0
propcache==0.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pyjnius==1.6.1
PyMySQL==1.1.1
pypiwin32==223
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.16
pywin32==308
reload==0.9
reportlab==4.2.5
requests==2.32.3
rich==13.9.3
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tkcalendar==1.6.1
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==5.0.3
yarl==1.17.0
